Split core etrade logic from cli and add a wasm api
