+++ src/peripheral.rs
@@ -7,6 +7,7 @@ pub mod excel;
 pub mod tx_export_convert_impl;
 
 #[cfg(feature = "pdf_parse")]
+pub mod etrade_plan_pdf_tx_extract_cmd;
 pub mod etrade_plan_pdf_tx_extract_impl;
 #[cfg(feature = "pdf_parse")]
 pub mod pdf;
